Fix signed-shift UB when building player/team bitmasks
Team::teamNumberToMask and BasePlayer::setNumber/setTeamNumber built
32-bit masks with a signed 1<<n; player and team indices can legally
reach 31 (Team::MAX_COUNT is 32), and 1<<31 on signed int is undefined
behavior. Shift an unsigned literal instead, route setTeamNumber
through the (now-fixed) Team::teamNumberToMask helper, and document
the helper's valid input range. No behavior change on real targets:
the G2.game replay stays byte-identical to the cpp-refactor baseline.

+++ src/team/Team.h
@@ -119,7 +119,10 @@ public:
 	void setCorrectMasks(void);
 	void setCorrectColor(const GAGCore::Color& color);
 	void setCorrectColor(float value);
-	inline static Uint32 teamNumberToMask(int team) { return 1<<team; }
+	/// Bit for team `team` in team-mask bitfields (Case::forbidden, discovery
+	/// masks, alliance masks, ...). Valid range: [0, MAX_COUNT), i.e. up to 31.
+	/// The unsigned one matters: `1<<31` on signed int is UB.
+	inline static Uint32 teamNumberToMask(int team) { return Uint32(1)<<team; }
 
 	void update();
 	bool openMarket();
